feat(barathonienController):refactoring

+++ app/Http/Controllers/BarathonienController.php
@@ -33,6 +33,10 @@ class BarathonienController extends Controller
                 })
                 ->get();
 
+            if ($barathoniens->isEmpty()) {
+                return $this->error(null, "No barathonien found", 404);
+            }
+
             return $this->success($barathoniens, "Barathonien List");
 
         } catch (Exception $error) {
@@ -49,47 +53,54 @@ class BarathonienController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
-        $today = new Carbon();
-        $minor = $today->subYears(18);
+        try {
 
-        $request->validate([
-            'first_name' => 'required|string|max:255',
-            'last_name' => 'required|string|max:255',
-            'email' => 'required|string|email|unique:users',
-            'password' => ['required', 'confirmed', Rules\Password::defaults()],
-            'birthday' => 'required|date|before:' . $minor,
-            'address' => 'min:5|required|string|max:255',
-            'postal_code' => 'required|string|size:5',
-            'city' => 'required|string|max:255',
-        ]);
+            $today = new Carbon();
+            $minor = $today->subYears(18);
 
-        $user = User::create([
-            'first_name' => $request->first_name,
-            'last_name' => $request->last_name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'avatar' => "https://picsum.photos/180",
-        ]);
+            $request->validate([
+                'first_name' => 'required|string|max:255',
+                'last_name' => 'required|string|max:255',
+                'email' => 'required|string|email|unique:users',
+                'password' => ['required', 'confirmed', Rules\Password::defaults()],
+                'birthday' => 'required|date|before:' . $minor,
+                'address' => 'min:5|required|string|max:255',
+                'postal_code' => 'required|string|size:5',
+                'city' => 'required|string|max:255',
+            ]);
 
-        $address = Address::create([
-            'address' => $request->address,
-            'postal_code' => $request->postal_code,
-            'city' => $request->city
-        ]);
+            $user = User::create([
+                'first_name' => $request->first_name,
+                'last_name' => $request->last_name,
+                'email' => $request->email,
+                'password' => Hash::make($request->password),
+                'avatar' => "https://picsum.photos/180",
+            ]);
 
-        $barathonien = Barathonien::create([
-            'birthday' => $request->birthday,
-            'address_id' => $address->address_id
-        ]);
+            $address = Address::create([
+                'address' => $request->address,
+                'postal_code' => $request->postal_code,
+                'city' => $request->city
+            ]);
+
+            $barathonien = Barathonien::create([
+                'birthday' => $request->birthday,
+                'address_id' => $address->address_id
+            ]);
 
 
-        $user->barathonien_id = $barathonien->barathonien_id;
-        $user->save();
+            $user->barathonien_id = $barathonien->barathonien_id;
+            $user->save();
 
-        return $this->success([
-            'userLogged' => $user,
-            'token' => $user->createToken('API Token')->plainTextToken
-        ], "Barathonien Created");
+            return $this->success([
+                'userLogged' => $user,
+                'token' => $user->createToken('API Token')->plainTextToken
+            ], "Barathonien Created", 201);
+
+        } catch (Exception $error) {
+            Log::error($error);
+            return $this->error(null, $error->getMessage(), 500);
+        }
     }
 
     /**
@@ -101,12 +112,14 @@ class BarathonienController extends Controller
     public function show($user_id): JsonResponse
     {
         try {
-            $barathonien = DB::table('users')
-                ->join('barathoniens', 'users.barathonien_id', '=', 'barathoniens.barathonien_id')
-                ->join('addresses', 'barathoniens.address_id', '=', 'addresses.address_id')
-                ->select('users.*', 'barathoniens.*', 'addresses.*')
-                ->where('users.user_id', '=', $user_id)
-                ->first();
+            $barathonien = User::with(['barathonien', 'barathonien.address'])
+                ->whereHas('barathonien', function ($query) use ($user_id) {
+                    $query->where('users.user_id', $user_id);
+                })
+                ->get();
+
+            if ($barathonien->isEmpty())
+                return $this->error(null, "Barathonien not found", 404);
 
             return $this->success($barathonien, "Barathonien");
 
@@ -129,6 +142,10 @@ class BarathonienController extends Controller
             // Get the user given in parameter
             $user = User::findOrFail($user_id);
 
+            // Check if the user is a barathonien
+            if ($user->barathonien_id === null)
+                return $this->error(null, "Barathonien not found", 404);
+
             $request->validate([
                 'first_name' => 'required|string|max:255',
                 'last_name' => 'required|string|max:255',
@@ -144,6 +161,7 @@ class BarathonienController extends Controller
             ]);
 
             $dataBarathonien = $request->only(['first_name', 'last_name', 'email']);
+
             // Check if the data given in parameter are different from the data in database
             foreach ($dataBarathonien as $field => $value) {
                 if ($user->{$field} !== $value) {
@@ -159,6 +177,7 @@ class BarathonienController extends Controller
             $address = Address::where('address_id', $barathonien->address_id)->first();
 
             $dataAddress = $request->only(['address', 'postal_code', 'city']);
+
             // Check if the data given in parameter are different from the data in database
             foreach ($dataAddress as $field => $value) {
                 if ($address->{$field} !== $value) {
@@ -166,6 +185,15 @@ class BarathonienController extends Controller
                 }
             }
             $address->save();
+
+            $userChanges = $user->getChanges();
+            $addressChanges = $address->getChanges();
+
+
+            // Check if the user has changed
+           // if (empty($userChanges) && empty($addressChanges))
+             //   return $this->success([$user, $address], "Barathonien not updated");
+
 
             // Return the updated user and address
             return $this->success([$user, $address], "Barathonien Updated");
@@ -177,7 +205,7 @@ class BarathonienController extends Controller
     }
 
     /**
-     * Deleting the barathonien
+     * Deleting the barathonien ( softDelete )
      *
      * @param $user_id
      * @return JsonResponse
@@ -185,8 +213,55 @@ class BarathonienController extends Controller
     public function destroy($user_id): JsonResponse
     {
         try {
+
+            //check if the user exist
+            $user = User::withTrashed()->where('user_id', $user_id)->first();
+            if ($user === null)
+                return $this->error(null, "User not found", 404);
+
+            //check if the user is a barathonien
+            if ($user->barathonien_id === null)
+                return $this->error(null, "Barathonien not found", 404);
+
+            //check if the user is already deleted
+            if ($user->deleted_at !== null)
+                return $this->error(null, "Barathonien already deleted", 404);
+
+            //delete the user
             User::where('user_id', $user_id)->delete();
-            return $this->success("", "Barathonien Deleted");
+            return $this->success(null, "Barathonien Deleted");
+
+        } catch (Exception $error) {
+            Log::error($error);
+            return $this->error(null, $error->getMessage(), 500);
+        }
+    }
+
+    /**
+     * Restoring the barathonien
+     *
+     * @param $user_id
+     * @return JsonResponse
+     */
+    public function restore($user_id): JsonResponse
+    {
+        try {
+            //check if the user exist
+            $user = User::withTrashed()->where('user_id', $user_id)->first();
+            if ($user === null)
+                return $this->error(null, "User not found", 404);
+
+            //check if the user is a barathonien
+            if ($user->barathonien_id === null)
+                return $this->error(null, "Barathonien not found", 404);
+
+            //check if the user is already restored
+            if ($user->deleted_at === null)
+                return $this->error(null, "Barathonien already restored", 404);
+
+            User::withTrashed()->where('user_id', $user_id)->restore();
+            return $this->success(null, "Barathonien Restored");
+
         } catch (Exception $error) {
             Log::error($error);
             return $this->error(null, $error->getMessage(), 500);
